# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: 匿名树洞社区 API
  version: 1.0.0
  description: 匿名发帖/评论、话题标签、点赞热度、每日精选、敏感词审核
servers:
  - url: http://localhost:19401
paths:
...
  /api/v1/posts:
    get:
      summary: 最新/标签/精选帖子列表
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: page_size
          schema: { type: integer }
        - in: query
          name: tag_id
          schema: { type: integer }
        - in: query
          name: featured
          schema: { type: boolean }
      responses:
        '200':
          description: 分页帖子列表
    post:
      summary: 发布帖子
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                title: { type: string }
                content: { type: string }
                images: { type: array, items: { type: string } }
                tags: { type: array, items: { type: string } }
      responses:
        '200':
          description: 帖子与命中信息
...
  /api/v1/posts/{id}:
    get:
      summary: 帖子详情（含楼主追记；楼主本人可见待审/已驳回追记及审核原因）
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 帖子详情
  /api/v1/posts/{id}/supplements:
    post:
      summary: 楼主追加追记（最多 2 条有效追记，每条不超过 500 字，提交后不可修改；命中敏感词进入审核；同一帖子仅允许一条待审追记，超限或并发冲突整次拒绝）
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string, maxLength: 500 }
      responses:
        '200':
          description: 追记与命中信息（blocked=true 表示已进入审核）
        '403':
          description: 非楼主
        '409':
          description: 追记已达上限 / 已有待审追记 / 帖子未发布
  /api/v1/posts/{id}/comments:
    get:
      summary: 帖子评论列表
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: page_size
          schema: { type: integer }
      responses:
        '200':
          description: 分页评论列表
  /api/v1/comments:
    post:
      summary: 发表评论
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [postId, content]
              properties:
                postId: { type: integer }
                content: { type: string }
      responses:
        '200':
          description: 评论与命中信息
...